% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterClass.R
\name{initialize_raw_parameter_class}
\alias{initialize_raw_parameter_class}
\title{initialize_raw_parameter_class}
\usage{
initialize_raw_parameter_class(column = c("hilic", "c18", "transponson"))
}
\arguments{
\item{column}{'hilic', 'c18'}
}
\value{
a object of RawMsParameterClass, with slots:
 \describe{
 \item{name}-{parameter set name}
 \item{para_peak_detection}-{parameter of peak detection}
 \item{para_peak_detection - method}-{peak detection method. Default: "centWave".}
 \item{para_peak_detection - method}-{peak detection method. Default: "centWave".}
 }
}
\description{
generate a RawMsParameterClass for processing raw mass spec data
}
\examples{
object <- initialize_raw_parameter_class(column = 'hilic')
}
\author{
Zhiwei Zhou
}
