% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionMain.R
\name{process_raw_data}
\alias{process_raw_data}
\title{process_raw_data}
\usage{
process_raw_data(parameter_set = NULL, is_skip_rtcor = FALSE, path = ".")
}
\arguments{
\item{parameter_set}{a parameter set class object (RawMsParameterClass). It can be generated easily by running \code{initialize_raw_parameter_class()}}

\item{path}{the path of raw MS data files}
}
\description{
a R wrapper to run xcms for raw MS data processing
}
\examples{
parameter_set <- initialize_raw_parameter_class(column = 'hilic')
process_raw_data(parameter_set = parameter_set, path = '~/Project/00_IBD_project/Data/20230327_raw_data_processing_test/DemoData/')
}
\author{
Zhiwei Zhou
}
